Add defense against a currently common worm
This change adds a new regex which protects the Gamers React server from a currently common worm. The worm takes over a user's Discord account to send download links to itself in all DMs and channels the user has access to. Essentially, it's spread by a message claiming it's a closed beta of a new game.